Handle more error cases including returning proper values from deduplicator

+++ dear.sh
@@ -8,32 +8,48 @@ function clean_up() {
     rm -rf /tmp/$$
 }
 
-# Show an error message if an unsupported number of commands is given
-if [ "$#" != 3 ] && [ "$#" != 2 ]; then
+function usage() {
     echo "Usage:"
     echo "./dear.sh -[compression] outfile indir"
     echo "Compression options:"
     echo "compress:  -c "
     echo "bzip:      -b "
     echo "gzip:      -g "
-    exit
+    exit 1
+}
+# Show an error message if an unsupported number of commands is given
+if [ "$#" != 3 ] && [ "$#" != 2 ]; then
+    usage;
 fi
 
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+# This mess is to set up the variables we need from the input
+# while also check for errors
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 if [ "$#" == 2 ]; then
     # if no switch is given use default extension
-    OUTFILE=$(basename "$1")    # output file name
-    OUTDIR=$(dirname "$1")/     # output file directory
-    INDIR=$2                    # input directory
-    BASEDIR=$(basename $2)      # get base directory
-    COMPRESS_SWITCH=""          # compression option
-    EXTENSION=".tar"            # extension of the compressed file
-    COMPRESS_TYPE="tar"         # Compression Type
+    OUTFILE=$(basename "$1")        # output file name
+    if [ $? -ne 0 ]; then usage; fi # Check to make sure success
+
+    OUTDIR=$(dirname "$1")/         # output file directory
+    INDIR=${2%/}                     # input directory
+
+    BASEDIR=$(basename $2)          # get base directory
+    if [ $? -ne 0 ]; then usage; fi # Check to make sure success
+
+    COMPRESS_SWITCH=""              # compression option
+    EXTENSION=".tar"                # extension of the compressed file
+    COMPRESS_TYPE="tar"             # Compression Type
 else
     # switch is given, determine the correct extension and compression method
     OUTFILE=$(basename "$2")
+    if [ $? -ne 0 ]; then usage; fi # Check to make sure success
+
     OUTDIR=$(dirname "$2")/
-    INDIR=$3
+
+    INDIR=${3%/}                    # input directory
     BASEDIR=$(basename $3)
+    if [ $? -ne 0 ]; then usage; fi # Check to make sure success
 
     if [ $1 != "-c" ] && [ $1 != "-b" ] && [ $1 != "-g" ]; then
         echo "Invalid compression option"
@@ -67,25 +83,30 @@ fi
 
 if [ ! -d ${OUTDIR} ]; then
     echo "cannot create outputfile ${OUTDIR}${OUTFILE}"
-    exit
+    exit 1
 fi
 
 # Test to make sure tha the input directory exists
 if [ ! -d ${INDIR} ]; then
     echo "Directory ${INDIR} does not exist"
-    exit
+    exit 1
 fi
 
 # Make sure output file doesn't already exist
 if [ -f ${OUTDIR}${OUTFILE}${EXTENSION} ]; then
     echo "File already exists ${OUTDIR}${OUTFILE}${EXTENSION}"
-    exit
+    exit 1
 fi
 
 # Set up temp directory, Is 2 directories deep so there are no collisions when moving compressed file
 # into the directory below it
 TMPDIR=/tmp/$$/$$
 CURRDIR=`pwd`
+
+if [ -d /tmp/$$ ]; then
+    echo "Could not create temporary directory"
+    exit 1
+fi
 
 mkdir -p ${TMPDIR}
 
@@ -95,7 +116,7 @@ cp -rf ${INDIR} ${TMPDIR}
 if [ $? -ne 0 ]; then
     echo "Couldn't copy files to temporary directory"
     clean_up
-    exit
+    exit 1
 fi
 # Copy duplicate find script, this is so that directory paths can remain relative
 cp ./dupRemove.pl ${TMPDIR}
@@ -103,6 +124,13 @@ cp ./dupRemove.pl ${TMPDIR}
 # Move into the temp directory, run deduplication script and compress
 cd ${TMPDIR}
 ./dupRemove.pl ./${BASEDIR}
+
+# Check to make deduplication succeeded
+if [ $? -ne 0 ]; then
+    echo "Deduplication failed"
+    exit 1
+fi
+
 rm ./dupRemove.pl
 
 # Place compressed file one directory up so that it isn't included in the archive
@@ -112,7 +140,7 @@ tar ${COMPRESS_SWITCH} -cf ../${OUTFILE}${EXTENSION} .
 if [ $? -ne 0 ]; then
     echo "Creating tar failed"
     clean_up
-    exit
+    exit 1
 else
     cd ${CURRDIR}
